eliminate the utterly broken "unescaped <" warning
It's not going to work, because the content that is in the paragraph
has already been E<>-decoded.  *Everything* is unescaped.  This needs
to be caught during parsing, I think.  I don't think that the para_end
event can have enough information.

I think we're okay without this warning for now, but it would be nice
to put it back.  In the meantime, rip it out.  The change to
podchkinter is to replace that "low severity warning" with another
one, since the test is for the behavior of -warnings, not the <
warning itself.