stop passing on Pod::Hyperlink warnings as warnings
This is part of the work to drop Pod::Hyperlink entirely.  Note that
this also drops a warning for L</I/O Foo>, which is no longer
ambiguous, and was already a buggy check.

A warning for this may be re-introduced via Pod::Simple in the
future, but probably not until Pod::Simple's L<> processing is
made less complex.